Re-org code to align with Ara's implementation
