+++ test/mongoid_store_test.rb
@@ -4,7 +4,7 @@ require "caching_test"
 class MongoidStoreTest < ActiveSupport::TestCase
   def setup
     @cache = ActiveSupport::Cache.lookup_store(:mongoid_store)
-    ::ActiveSupport::Cache::MongoidStore::Entry.delete_all
+    ::ActiveSupport::Cache::MongoidStore::Storage.delete_all
   end
 
   include ::CacheStoreBehavior
